Refactor authentication and diary components; remove unused LayoutClient files and update type imports
