+++ app/(auth routes)/auth/AuthForm.module.css
@@ -1,42 +1,20 @@
-/* .authContainer {
+.formWrapper {
   display: flex;
+  justify-content: center;
+  align-items: center;
   flex-direction: column;
   min-height: 100vh;
-}
-
-.sideImageWrapper {
-  flex: 1;
-  display: none;
-  align-items: center;
-  justify-content: center;
-}
-
-.sideImageWrapper img {
-  max-width: 100%;
-  height: auto;
+  margin: 0 auto;
 }
 
 .form {
   display: flex;
   flex-direction: column;
   gap: 16px;
-  max-width: 335px;
+  width: 335px;
   margin: 0 auto;
   padding: 20px;
   border: none;
-  max-width: 335px;
-  width: 100%;
-  margin: 0 auto;
-  padding: 20px;
-  border: none;
-}
-.formWrapper {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  min-height: 100vh;
 }
 
 .fieldset {
@@ -59,7 +37,6 @@
 .label {
   display: block;
   margin-bottom: 8px;
-  font-family: var(--font-family);
   font-weight: 400;
   font-size: 14px;
   line-height: 150%;
@@ -74,7 +51,6 @@
   padding: 8px 12px;
   height: 37px;
   background: var(--opacity-neutral-darkest-5);
-  font-family: var(--font-family);
   font-weight: 400;
   font-size: 14px;
   line-height: 150%;
@@ -84,18 +60,19 @@
 .inputError {
   border: 1px solid var(--color-red);
   border-radius: 12px;
-  font-family: var(--font-family);
   font-weight: 400;
   font-size: 16px;
   line-height: 150%;
   color: var(--color-red);
+  margin-bottom: 8px;
 }
+
 .error {
-  font-family: var(--font-family);
   font-weight: 400;
   font-size: 12px;
   line-height: 160%;
   color: var(--color-red);
+  margin-bottom: 8px;
 }
 
 .submitButton,
@@ -108,10 +85,7 @@
   width: 100%;
   height: 42px;
   cursor: pointer;
-  font-family: var(--font-family);
-  font-weight: 500;
   font-size: 14px;
-  line-height: 160%;
   color: var(--color-scheme-text);
 }
 
@@ -127,20 +101,13 @@
 
 .loginLink,
 .registerLink {
-  font-family: var(--font-family);
   font-weight: 400;
   font-size: 14px;
   line-height: 150%;
   text-align: center;
-  color: var(--color-scheme-text);
 }
 
 .link {
-  font-family: var(--font-family);
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 160%;
-  text-align: center;
   color: var(--color-scheme-text);
   text-decoration: none;
 }
@@ -149,217 +116,40 @@
   text-decoration: underline;
 }
 
+.imageWrapper {
+  display: none;
+}
+
 @media (min-width: 768px) {
   .form {
-    max-width: 480px;
+    width: 480px;
   }
 }
 
-@media screen and (min-width: 1024px) {
+@media (min-width: 1440px) {
+  .formWrapper {
+    flex: 1;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .form {
+    width: 100%;
+    max-width: 480px;
+    margin: 0;
+  }
+
   .authContainer {
     display: flex;
-    flex-direction: row;
-  }
-
-  .formWrapper {
-    display: flex;
-    flex: 1;
-    align-items: center;
-    justify-content: center;
-    min-height: 100vh;
-  }
-
-  .form {
-
+    justify-content: space-between;
     width: 100%;
   }
 
-  .sideImageWrapper {
+  .imageWrapper {
     position: relative;
-    display: block;
-    flex: 1;
-    width: 100vh;
+    width: 100%;
     height: 100vh;
-  }
-} */
-
-.authContainer {
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-}
-
-.formWrapper {
     flex: 1;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    min-height: 100vh;
-    margin: 0 auto;
-}
-
-.form {
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-    width: 335px;
-    /* Mobile width */
-margin: 0 auto;
-    padding: 20px;
-    border: none;
-}
-
-.fieldset {
-    border: none;
-    padding: 0;
-    margin: 0;
-}
-
-.legend {
-    font-family: var(--second-family);
-    font-weight: 700;
-    font-size: 36px;
-    line-height: 120%;
-    letter-spacing: 0.01em;
-    text-align: center;
-    padding-bottom: 32px;
-    color: var(--color-scheme-text);
-}
-
-.label {
     display: block;
-    margin-bottom: 8px;
-    font-family: var(--font-family);
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 150%;
-    color: var(--color-scheme-text);
-}
-
-.input {
-    width: 100%;
-    margin-bottom: 24px;
-    border: 2px solid var(--opacity-transparent);
-    border-radius: 12px;
-    padding: 8px 12px;
-    height: 37px;
-    background: var(--opacity-neutral-darkest-5);
-    font-family: var(--font-family);
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 150%;
-    color: var(--opacity-neutral-darkest-60);
-}
-
-.inputError {
-    border: 1px solid var(--color-red);
-    border-radius: 12px;
-    font-family: var(--font-family);
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 150%;
-    color: var(--color-red);
-}
-
-.error {
-    font-family: var(--font-family);
-    font-weight: 400;
-    font-size: 12px;
-    line-height: 160%;
-    color: var(--color-red);
-}
-
-.submitButton,
-.button {
-    border-bottom: 4px solid var(--color-scheme-accent);
-    background: var(--color-scheme-foreground-alt);
-    border-radius: 100px;
-    border: none;
-    padding: 10px 24px;
-    width: 100%;
-    height: 42px;
-    cursor: pointer;
-    font-family: var(--font-family);
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 160%;
-    color: var(--color-scheme-text);
-}
-
-.submitButton:hover,
-.button:hover {
-    background: var(--color-scheme-foreground);
-}
-
-.submitButton:active,
-.button:active {
-    background: var(--color-scheme-accent);
-}
-
-.loginLink,
-.registerLink {
-    font-family: var(--font-family);
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 150%;
-    text-align: center;
-    color: var(--color-scheme-text);
-}
-
-.link {
-    font-family: var(--font-family);
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 160%;
-    text-align: center;
-    color: var(--color-scheme-text);
-    text-decoration: none;
-}
-
-.link:hover {
-    text-decoration: underline;
-}
-
-.sideImageWrapper {
-    display: none;
-    /* Hidden by default on mobile */
-}
-
-@media (min-width: 768px) {
-    .form {
-        width: 480px;
-        /* Tablet width */
-    }
-}
-
-@media (min-width: 1024px) {
-    .authContainer {
-        flex-direction: row;
-    }
-
-    .formWrapper {
-        flex: 0 0 50%;
-        /* Form takes the left half */
-        justify-content: center;
-        align-items: center;
-    }
-
-    .form {
-        width: 100%;
-        /* Form fills its 50% container */
-        max-width: 480px;
-        /* Or a fixed width if preferred within the half */
-        margin: 0;
-        /* Remove auto margin to align within the half */
-    }
-
-    .sideImageWrapper {
-        display: block;
-        /* Show image on desktop */
-        flex: 0 0 50%;
-        /* Image takes the right half */
-        position: relative;
-        height: 100vh;
-    }
+  }
 }
