+++ app/(private routes)/diary/[entryId]/DiaryEntryClientPage.tsx
@@ -2,7 +2,7 @@
 
 import DiaryEntryDetails from '@/components/DiaryEntryDetails/DiaryEntryDetails';
 import { getDiariesById } from '@/lib/api';
-import { DiaryData } from '@/lib/types';
+import { DiaryData } from '@/types/types';
 import { useQuery } from '@tanstack/react-query';
 import { PuffLoader } from 'react-spinners';
 import css from './DiaryEntryClientPage.module.css';
