+++ app/(private routes)/layout.tsx
@@ -1,0 +1,9 @@
+import LayoutClient from '@/components/LayoutClient/LayoutClient';
+
+export default function PrivateLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <LayoutClient>{children}</LayoutClient>;
+}
