+++ app/layout.tsx
@@ -2,7 +2,6 @@ import type { Metadata } from 'next';
 import './globals.css';
 import { Lato, Comfortaa } from 'next/font/google';
 import TanStackProvider from '@/components/TanStackProvider/TanStackProvider';
-import LayoutClient from './LayoutClient';
 
 const lato = Lato({
   subsets: ['latin'],
@@ -29,11 +28,8 @@ export default function RootLayout({
   return (
     <html lang="uk">
       <body className={`${lato.variable} ${comfortaa.variable}`}>
-        <TanStackProvider>
-          <LayoutClient>{children}</LayoutClient>
-        </TanStackProvider>
+        <TanStackProvider>{children}</TanStackProvider>
       </body>
     </html>
   );
 }
-
