+++ app/page.tsx
@@ -1,3 +1,9 @@
-export default function DashboardPage() {
-  return;
+import LayoutClient from '../components/LayoutClient/LayoutClient';
+
+export default function DashboardPage({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <LayoutClient>{children}</LayoutClient>;
 }
