+++ components/Sidebar/Sidebar.module.css
@@ -22,6 +22,7 @@
   justify-content: space-between;
   align-items: center;
   padding: 16px 24px;
+  padding-bottom: 0;
 }
 
 .btn {
@@ -161,6 +162,7 @@
 
   .logo_container {
     padding: 17px 28px;
+    padding-bottom: 0;
   }
 
   .link {
