+++ lib/store/diaryStore.ts
@@ -1,12 +1,12 @@
 import { create } from 'zustand';
-import { DiaryData } from '@/lib/types';
+import { DiaryData } from '@/types/types';
 
 type DiaryState = {
   selectedDiary: DiaryData | null;
   setSelectedDiary: (diary: DiaryData | null) => void;
 };
 
-export const useDiaryStore = create<DiaryState>((set) => ({
+export const useDiaryStore = create<DiaryState>(set => ({
   selectedDiary: null,
-  setSelectedDiary: (diary) => set({ selectedDiary: diary }),
+  setSelectedDiary: diary => set({ selectedDiary: diary }),
 }));
